Split GraphicsCapabilities into partial class for overloading on console platforms.
